routes_app/pyproject.toml
[tool.poetry]
name = "routes-app"
version = "1.0.0"
description = "API REST para gestión de rutas/trayectos"
authors = ["jorge"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.111"
uvicorn = {extras = ["standard"], version = ">=0.30"}
psycopg2-binary = ">=2.9"
pydantic = ">=2.6"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2"
pytest-cov = ">=5.0"
ruff = ">=0.3"
black = ">=24.0"
httpx = ">=0.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --disable-warnings"

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
fail_under = 70
show_missing = true
